[REFACTOR #46] 수리 요청 리팩토링
