+++ src/main/java/com/duktown/domain/repairApply/dto/RepairApplyDto.java
@@ -24,7 +24,7 @@ public class RepairApplyDto {
         @Max(value = 2, message = "카테고리는 0에서 2사이의 정수 값입니다.")
         private Integer hallName;
 
-        private String unit;
+        private String roomNumber;
 
         @NotBlank(message = "내용은 필수 입력값입니다")
         private String content;
@@ -34,7 +34,7 @@ public class RepairApplyDto {
             return RepairApply.builder()
                     .user(user)
                     .hallName(hallName)
-                    .unit(unit)
+                    .roomNumber(roomNumber)
                     .content(content)
                     .checked(false)
                     .solved(false)
@@ -49,7 +49,7 @@ public class RepairApplyDto {
         private Long id;
         private Long userId;
         private Integer hallName;
-        private String unit;
+        private String roomNumber;
         private String content;
         private Boolean checked;
         private Boolean solved;
@@ -59,7 +59,7 @@ public class RepairApplyDto {
             this.id = apply.getId();
             this.userId = apply.getUser().getId();
             this.hallName = apply.getHallName().getValue();
-            this.unit = apply.getUnit();
+            this.roomNumber = apply.getRoomNumber();
             this.content = apply.getContent();
             this.checked = apply.getChecked();
             this.solved = apply.getSolved();
